Modified Weather features

+++ InterviewTask/Models/OpenWeatherMapServiceModel.cs
@@ -21,10 +21,10 @@ namespace InterviewTask.Models
         public class Main
         {
             public double temp { get; set; }
-            public int pressure { get; set; }
-            public int humidity { get; set; }
-            public int temp_min { get; set; }
-            public int temp_max { get; set; }
+            public double pressure { get; set; }
+            public double humidity { get; set; }
+            public double temp_min { get; set; }
+            public double temp_max { get; set; }
         }
 
         public class Wind
